Address Guillaume second review
- Rename MortarData -> MortarInterfaceWarehouse
- More descriptive wording in AutomaticMortarGeneration
- Better method parameter names
