+++ framework/src/interfaces/MortarConsumerInterface.C
@@ -13,7 +13,7 @@
 #include "MooseObject.h"
 #include "FEProblemBase.h"
 #include "MooseMesh.h"
-#include "MortarData.h"
+#include "MortarInterfaceWarehouse.h"
 #include "Assembly.h"
 #include "AutomaticMortarGeneration.h"
 #include "libmesh/quadrature.h"
